chore(poly): modularized
**Motivation**
Modifications to this module are expected (#159), increased modularity
will simplify component modification and isolation.

**Overview**
Not a single line of code was essentially produced, the current code was
stripped into themed modules without modification
